feat: road expansion v2 — double-width ribbons + grass medians, and the city became navigable
Widths doubled on all four classes (spine 22.0 / artery 19.8 / major 17.6 / minor
13.2 wu, still CAR_REF-derived, zero literals) with a 2.2 wu visual-only raised
grass median on the spine + three arteries (per-street optIn wired for major,
unused) and 136 sparse median trees whose pitch is DERIVED from the camera's
visible ground band, so at most one is ever in frame.

NAVIGABILITY, vs Phase 74's own high-tier baseline (same routes, same seed):
downtownDense mean 1.6→3.9 m/s (path 92→200 m), spineCruise 8.2→13.4 (489→791 m),
minorWeave 3.4→6.9 (195→397 m); unrecoverable stuck events 2+2→0 everywhere; and
minorWeave's 30.7% airtime / 3 flips / 1.57 rad roll → 0 / 0 / 0.04. Contacts/min
rose 14→27 but contacts PER METRE fell 0.153→0.135 — more driving, not more
crashing.

LANE_OFFSET_WU re-derived to the carriageway centre with the 2.2 cap deleted (it
would have bound on every class, straddling the median while 8+ wu of new asphalt
sat unused), replaced by a lane-inside-ribbon/outside-median law. Asphalt now emits
as a class-ordered UNION — Bay/York overlap by 7.88 wu and Bay's half-width alone
exceeds their 7.52 wu gap, so no width override can separate them (they were
already tangent at -0.18 wu pre-phase: a 0.6-compaction artifact merely exposed).
The union also removed the pre-existing coplanar pair at all 73 intersections.
New law: no two lane chains anywhere within one car width — Bay/York had put
opposing chains 0.184 wu apart, Phase 31's head-on deadlock reborn; fixed via
swallowedSpans, worst cross-street separation now 25.5 wu (11.6x the floor).

Junction rebuild: CROSSWALK.bandWu 2.2→3.0 (P27 cut it because the roads were
dieted; doubling inverts that exactly), traffic-light scale 1.0→1.74 with the mast
law made two-sided (it bounded reach only from above and passed trivially at 2x
width), lamp anchor and post claim re-derived, and the mast collider moved off the
model's full bounding box to the post half-width — killing ~3.6 wu of invisible
diagonal wall at every signalized corner.

Four landmark throw-guards fired for real. Osgoode Hall root-fixed by generalizing
Frontage to accept a corner pair, and three named buildings migrated onto it
(HHOF 3.50 / Royal York 2.73 / Old City Hall 2.30 wu into ribbons) — all the same
bug: a facade gap chosen off a ribbon edge but written as a centreline offset.

FLICKER: one hotspot found and proven new against a pre-phase worktree, then
root-caused to a curb strip losing its depth test to its own asphalt (a 0.006 wu
decal = ~4-8 depth-buffer LSBs). Latent since long before this phase (7% spread
before, 40.5% after) — widening pushed it past threshold. Fixed by folding curb
strips into the ribbon union at ROAD_Y, no epsilon touched: 1.2% spread, hot tiles
18 (pre-phase) / 258 (broken) → 11, better than baseline. Two harness debts closed:
flicker EXPECTED_VANTAGE_COUNT was stale at 183 so every run exited 1 even at zero
hotspots, and money-dash-far had lost its subject (a curated dash-study anchor
sitting on the now-median spine) — re-targeted onto Eglinton.

Land trade measured, not asserted: road area +93.6%, buildable land -7.70%,
frontage length -0.13%, shipped streetwall flat at the cap but raw supply -10.8%
(re-pinned as a measurement + tripwire). polygon.ts's stated area was a stale
pre-compaction figure, wrong by 2.7x since Part-8, and had misled the plan —
corrected in-file.

Goldens re-pinned once, claims 8076→8182 fully attributed, arbiter sweep green
across 3 seeds with zero tolerances widened and zero new ribbon exceptions.

Gates: 4,131 tests (+161), bench x3 tiers (high 73/150 · 152.2k, med 72/120 ·
109.2k, low 66/90 · 97,968/120k — materially better than P74's 110.9k), flicker
4 slices / 94 vantages CLEAN, smoke 31/31, eyeInside 0.0% at all 7 vantages before
and after, draw calls DOWN (63→57, 63→58, 60→58), 0 console errors.

KNOWN SHORTFALL: turn radius is still not measurable. The corridor did open
(±8.5 → ±14.0 m clear) and @20 progressed from refusing before the car moved to
actually driving the manoeuvre, but a steady full-lock circle at 20 m/s is
grip-limited to a ~40 m radius and needs ~80 m of corridor — no street of any
width can hold it — and @5/@10 fail on settling and heuristics unrelated to
geometry. The probe's design, not the map, is the binding constraint; the quantity
is already available from stepSteer (R = v/w = 8.56 m at 15.5 m/s). Phase 78
should re-express the metric and re-pin the already-provisional Feel Spec rows.

+++ src/game/world/toronto/unionStation.test.ts
@@ -258,8 +258,13 @@ describe('Union Station v2 — the sunken carriageway (moat)', () => {
       maxZ: front.ribbon.maxY,
     };
     expect(overlaps(moat, ribbon)).toBe(false);
-    // …and no vertex of the whole mesh crosses the kerb either.
-    expect(Math.min(...verts.map((v) => v.z))).toBeGreaterThanOrEqual(moat.minZ - 1e-6);
+    // …and no vertex of the whole mesh crosses the kerb either. The bound is rounded to the
+    // precision the vertex actually has: `verts` is read back out of a Float32Array, and the moat's
+    // north row of vertices sits exactly ON `moat.minZ`, so the stored value is fround(minZ) —
+    // which at this coordinate's magnitude is up to 1.2e-4 wu away from the float64 original, i.e.
+    // ~240× the 1e-6 slack below. PHASE 75 exposed that (Front's kerb moved 4.4 wu with the road
+    // widening and the new value happens to round DOWN); the assertion itself is unchanged.
+    expect(Math.min(...verts.map((v) => v.z))).toBeGreaterThanOrEqual(Math.fround(moat.minZ) - 1e-6);
   });
 });
